fix(perps): finish the fee-copy fixes the first pass left half-done
Four issues from re-review, three of them created or missed by the previous
commit:

- The two-channel hint overflowed the info dialog. The Table is
  `table-fixed ... sm:whitespace-nowrap`, so the longer web/API string could
  not wrap and ran outside the modal. Shortened to "pool-sized: ~X web / ~Y
  API" and the cell now opts back into wrapping.

- The signed-out mobile fix was only half a fix. Hiding the "··· → See info"
  route left "is listed in the market info panel" — still a reference to
  something that reader cannot open, since ContractInfoDialog is reachable
  ONLY via HeaderActions and both of its containers are `!user && 'hidden
  md:flex'`. The always-visible sentence is now complete and true on its own
  ("set per market, and only Manifold admins can change them"), with the
  whole panel reference behind the same condition that renders the panel.

- The settings dialog still floored what the explainer now prints exactly.
  formatWithCommas is Math.floor, so a 2.5x cap read "2x" and an ACTIVE 0.5
  impact read "0" — and the explainer, fixed last commit, sat next to it
  saying 2.5x. Both reader and admin branches now use formatNumber at 2dp.

- apiDiffers compared raw bps while every consumer renders through
  formatFeePct, so base 10 vs API 10.1 turned on a second sentence whose
  every figure was character-identical to the first — the exact redundancy
  the flag exists to prevent. It now also requires the difference to survive
  formatting.

That last one means the summary has to know how its numbers are rendered, so
perpFeeScheduleSummary moves from fees.ts to format.ts (which already owns
formatFeePct). Better layering regardless: fees.ts is on the engine's import
path and should not pull in a formatter. Its tests move with it, plus two
pinning both sides of the new gate.
