+++ common/src/perps/format.ts
@@ -1,3 +1,10 @@
+import {
+  getPerpEffectiveTakerFeeBps,
+  getPerpTakerFeeBps,
+  getPerpTakerFeeImpact,
+  perpFreshPositionFeeBps,
+} from './fees'
+
 // Infer how many decimal places to show for an oracle-price-style value.
 // Heuristic: if every sample is integer-valued, show 0 decimals; otherwise
 // scale decimals to the value's magnitude so big prices don't look like
@@ -61,4 +68,80 @@ export const formatFeePct = (bps: number) => {
   if (Number(pct.toFixed(2)) < 1) return `${pct.toFixed(2)}%`
   if (Number(pct.toFixed(1)) < 10) return `${pct.toFixed(1)}%`
   return `${Math.round(pct)}%`
+}
+
+/** Pool shares the reader-facing fee examples are worked at. A pool-sized
+ * entry is the intuitive unit ("as big as the market backing it"); 4x is the
+ * calibration reference the impact default was chosen against (see
+ * PERP_TAKER_FEE_IMPACT_DEFAULT). Named so the explainer, the info dialog and
+ * that comment cannot drift to different examples. */
+export const PERP_FEE_EXAMPLE_POOL_SHARES = [1, 4] as const
+
+export type PerpFeeScheduleSummary = {
+  baseBps: number // web open rate
+  apiBps: number // API-key open rate (>= base; equals base when unset)
+  /** Whether the API channel is worth a sentence of its own — true only when
+   * it costs more AND that difference SURVIVES formatting. Comparing raw bps
+   * is not enough: base 10 vs API 10.1 is a real difference the engine
+   * charges, but every figure it produces renders identically at display
+   * precision, so saying it twice reads as a bug rather than as information.
+   * Gated here, not at each call site, so all surfaces agree on when to
+   * speak. */
+  apiDiffers: boolean
+  impact: number
+  hasSizeTerm: boolean
+  /** Effective rates for a FRESH position at each example pool share, per
+   * channel. The size term stacks on whichever base the CHANNEL selected —
+   * the engine picks the base first and then scales — so the web figures are
+   * simply wrong for an API open and both must be available to render. */
+  poolSizedBps: number
+  fourTimesPoolBps: number
+  apiPoolSizedBps: number
+  apiFourTimesPoolBps: number
+}
+
+/**
+ * A market's fee schedule reduced to the numbers a reader needs, derived in
+ * ONE place so the perp explainer and the market info dialog cannot quote the
+ * same market two different ways. Every figure routes through
+ * perpFreshPositionFeeBps -> perpSizeFeeDetails, i.e. the math the engine
+ * charges from.
+ *
+ * Lives here rather than in fees.ts because it is a DISPLAY reduction — it
+ * has to know how its numbers will be rendered (see apiDiffers) — and fees.ts
+ * is on the engine's import path, which should not pull in a formatter.
+ *
+ * Total, like the getters it composes: corrupt legacy config reads as the
+ * defaults rather than throwing inside a React render.
+ */
+export const perpFeeScheduleSummary = (contract: {
+  takerFeeBps?: number
+  takerFeeApiBps?: number
+  takerFeeImpact?: number
+}): PerpFeeScheduleSummary => {
+  const baseBps = getPerpTakerFeeBps(contract)
+  const apiBps = getPerpEffectiveTakerFeeBps(contract, true)
+  const impact = getPerpTakerFeeImpact(contract)
+  const [poolShare, bigShare] = PERP_FEE_EXAMPLE_POOL_SHARES
+  const at = (b: number, share: number) =>
+    perpFreshPositionFeeBps({ baseBps: b, impact, poolShare: share })
+  const poolSizedBps = at(baseBps, poolShare)
+  const fourTimesPoolBps = at(baseBps, bigShare)
+  const apiPoolSizedBps = at(apiBps, poolShare)
+  const apiFourTimesPoolBps = at(apiBps, bigShare)
+  const rendersIdentically =
+    formatFeePct(baseBps) === formatFeePct(apiBps) &&
+    formatFeePct(poolSizedBps) === formatFeePct(apiPoolSizedBps) &&
+    formatFeePct(fourTimesPoolBps) === formatFeePct(apiFourTimesPoolBps)
+  return {
+    baseBps,
+    apiBps,
+    apiDiffers: apiBps > baseBps && !rendersIdentically,
+    impact,
+    hasSizeTerm: impact > 0,
+    poolSizedBps,
+    fourTimesPoolBps,
+    apiPoolSizedBps,
+    apiFourTimesPoolBps,
+  }
 }
